# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the API server for Teralux App.",
        "title": "Teralux API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/cache/flush": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Remove all data from the cache storage",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "05. Flush"
                ],
                "summary": "Flush all cache",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtos.StandardResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dtos.StandardResponse"
                        }
                    }
                }
            }
        },
        "/api/tuya/auth": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Authenticates the user and retrieves a Tuya access token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "01. Auth"
                ],
                "summary": "Authenticate with Tuya",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dtos.StandardResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dtos.TuyaAuthResponseDTO"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dtos.StandardResponse"
                        }
                    }
                }
            }
        },
        "/api/tuya/devices": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves a list of all devices. Response format depends on GET_ALL_DEVICES_RESPONSE_TYPE: 0 (Nested/Default), 1 (Flat), 2 (Merged). Sorted alphabetically by Name. For infrared_ac devices, the status array is populated with saved device state (power, temp, mode, wind) or default values if no state exists.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "02. Devices"
                ],
                "summary": "Get All Devices",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Items per page",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by category",
                        "name": "category",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dtos.StandardResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dtos.TuyaDevicesResponseDTO"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dtos.StandardResponse"
                        }
                    }
                }
            }
        },
...
        "/api/tuya/devices/{id}/commands/ir": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Sends an infrared command to an AC via a specific IR device",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "03. Device Control"
                ],
                "summary": "Send IR AC Command",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Infrared Device ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "IR AC Command Payload",
                        "name": "command",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.TuyaIRACCommandDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtos.StandardResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dtos.StandardResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dtos.StandardResponse"
                        }
                    }
                }
            }
        },
        "/api/tuya/devices/{id}/commands/switch": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Sends a command to a specific Tuya device",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "03. Device Control"
                ],
                "summary": "Send Command to Device",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Device ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Command Payload",
                        "name": "command",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.TuyaCommandDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtos.StandardResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dtos.StandardResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dtos.StandardResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dtos.StandardResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean"
                }
            }
        },
        "dtos.TuyaAuthResponseDTO": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string"
                },
                "expire_time": {
                    "type": "integer"
                },
                "refresh_token": {
                    "type": "string"
                },
                "uid": {
                    "type": "string"
                }
            }
        },
        "dtos.TuyaCommandDTO": {
            "type": "object",
            "required": [
                "code",
                "value"
            ],
            "properties": {
                "code": {
                    "type": "string"
                },
                "value": {}
            }
        },
        "dtos.TuyaDeviceDTO": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "string"
                },
                "collections": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dtos.TuyaDeviceDTO"
                    }
                },
                "create_time": {
                    "type": "integer"
                },
                "custom_name": {
                    "type": "string"
                },
                "gateway_id": {
                    "type": "string"
                },
                "icon": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "ip": {
                    "type": "string"
                },
                "local_key": {
                    "type": "string"
                },
                "model": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "online": {
                    "type": "boolean"
                },
                "product_name": {
                    "type": "string"
                },
                "remote_category": {
                    "type": "string"
                },
                "remote_id": {
                    "type": "string"
                },
                "remote_product_name": {
                    "type": "string"
                },
                "status": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dtos.TuyaDeviceStatusDTO"
                    }
                },
                "update_time": {
                    "type": "integer"
                }
            }
        },
...
        "dtos.TuyaDeviceStatusDTO": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "value": {}
            }
        },
        "dtos.TuyaDevicesResponseDTO": {
            "type": "object",
            "properties": {
                "current_page_count": {
                    "type": "integer"
                },
                "devices": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dtos.TuyaDeviceDTO"
                    }
                },
                "total_devices": {
                    "type": "integer"
                }
            }
        },
        "dtos.TuyaIRACCommandDTO": {
            "type": "object",
            "required": [
                "code",
                "remote_id"
            ],
            "properties": {
                "code": {
                    "type": "string"
                },
                "remote_id": {
                    "type": "string"
                },
                "value": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "X-API-KEY",
            "in": "header"
        },
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}